Abstract team creation into bottle creation
Cohorts and teams were basically the same thing with different limits.
Now we have "things that hold stuff" called bottles.  A team is a
bottle with certain settings, and a cohort is a type of bottle with
certain settings.  Courses will also be bottles as well as self-reverse
course mappings.

All tests pass.